refactor: GiftOrderPage 리팩토링
- OrderTypes 파일 작성
- GiftOrderPage/hooks 폴더 작성
- GiftOrderPage 리팩토링
