+++ README.md
@@ -77,4 +77,9 @@
 - postRequest 작성
 - post 요청하는 페이지 수정
 
+#### GiftOrderPage 리팩토링
+
+- 기존 페이지는 역할이 너무 많고 복잡
+- 관심사 분리를 위해 훅을 만들어서 리팩토링
+
 ## 리뷰 반영
